Move tp tests to distributed tests and reduce cost
